About WebUI moved to its own directory.
Chromium change:

https://chromium.googlesource.com/chromium/src/+/5d34db1191f37

commit 5d34db1191f3791d55007d8989380b7ca8e3e1cb
Author: Rebekah Potter <rbpotter@chromium.org>
Date:   Mon Nov 13 21:24:21 2023 +0000

    About UI: Move to dedicated folder

    This is in preparation for making other updates to chrome://chrome-urls.

    The follow-up CL adding such updates also fixes the low test coverage
    reported in these files, see:
    https://chromium-review.googlesource.com/c/chromium/src/+/5012056

    Low-Coverage-Reason: LARGE_SCALE_REFACTOR
    Bug: 1500760

https://chromium.googlesource.com/chromium/src/+/184d907

commit 184d9073144c8a5d81a966b65b59472927fc16ac
Author: Rebekah Potter <rbpotter@chromium.org>
Date:   Tue Nov 14 20:38:11 2023 +0000

    WebUI: Add a way to list all registered WebUIs and use in chrome-urls

    Adding a GetWebUIConfigList() method to WebUIConfigMap that returns
    a vector of information about all registered WebUIConfigs. This
    includes the origin - host and scheme - and whether the UI is enabled
    or disabled.

    Leveraging this method to augment the list of URLs listed in
    chrome://chrome-urls, such that new WebUIs do not need to be manually
    added to this list if they are added using WebUIConfig.

    Also removing some dead untested ChromeOS code in about_ui.cc that was
    causing Code Coverage -1 blocking CL submission.

    Bug: 1500760
